// packages/pqb/src/query/pick-query-types.d.ts
export interface PickQueryReturnType {
    returnType: QueryReturnType;
}
// ...
export interface PickQueryWithData {
    withData: WithDataItems;
}
export interface PickQueryWindows {
    windows: EmptyObject;
}
export interface PickQueryRelations {
    relations: RelationsBase;
}
// ...
export interface PickQueryRelationsWithData extends PickQueryWithData, PickQueryRelations {
}
// ...
export interface PickQueryQ {
    q: QueryData;
}